requirements.txt
# Core ML Libraries
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Data Processing
numpy>=1.21.0
pandas>=1.5.0
scikit-learn>=1.2.0
opencv-python>=4.7.0
Pillow>=9.0.0
scipy>=1.9.0

# Visualization
matplotlib>=3.6.0
seaborn>=0.12.0
plotly>=5.10.0

# Medical Imaging
pydicom>=2.3.0
nibabel>=4.0.0
SimpleITK>=2.2.0

# Explainable AI
captum>=0.6.0

# Web Interface
streamlit>=1.20.0
gradio>=3.20.0

# Experiment Tracking
mlflow>=2.0.0
wandb>=0.13.0

# Utilities
tqdm>=4.64.0
pyyaml>=6.0
requests>=2.28.0
timm>=0.6.0  # For additional model architectures

# Development
pytest>=7.0.0
black>=22.0.0
flake8>=5.0.0
mypy>=1.0.0

# Optional: For advanced preprocessing
albumentations>=1.3.0
imbalanced-learn>=0.10.0